Enable RiakControl.LoadingRoute for Ember 1.0
This commit adds loading spinners to the various Riak Control
states and uses Ember's loadingRoute functionality where
possible.

LoadingRoute works by default on the `snapshot` and `nodes` pages
because, in the router, each state's `model` property returns an
instance of `this.store.find` which Ember interprets to involve a
`Promise` object. Since we have to return manually instantiated
objects in the `model` property for the `cluster` and `ring` routes,
we can't return a `Promise` object and thus have to manually
render the loading template for those two states.